new: SCRIPTS_AS_STATIC_DIRECTORY
Before, the default behavior was Eden would require every module regardless of where it was as long as it wasn't inside of a "static" folder.

Now with this configuration option, users have the ability to make Eden recognize script instances as static directories as well.

This is nice because now Eden wont require descendant modules of a module, as this would rarely be the desired behavior.